+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Add request env to sampling context when using `sentry-rails` [#1792](https://github.com/getsentry/sentry-ruby/pull/1792)
   - Fixes [#1791](https://github.com/getsentry/sentry-ruby/issues/1791)
 - Fix net-http tracing's span nesting issue [#1796](https://github.com/getsentry/sentry-ruby/pull/1796)
+- Don't mask real error when converting an event to JSON fails [#1797](https://github.com/getsentry/sentry-ruby/pull/1797)
 
 ### Refactoring